Fixed problems involving relative positioning for level buttons, now mouse position is converted to be relative to a specific widget when checking if something is clicked.
Option buttons can be detected for each level. Levels can now be deleted but the menu reset is quite jarring. I do not plan to fix this as it is not a priority.

I do want to improve scrolling for the menu if possible so a higher scroll speed can be used, but the priority now is to get sprites for the game and to start making a level creator.